requirements.txt
pytest==7.1.2
allure-pytest==2.13.2
allure-python-commons==2.13.2
requests==2.31.0
